feat: refactor ApplicationSettingsPage to use ApplicationGeneralSettingsView and ApplicationRestrictionsView components
