+++ src/packages/ce/src/application/pages/ApplicationSettingsPage.tsx
@@ -1,26 +1,10 @@
 "use client"
 
-import React, {startTransition} from "react";
-import {
-    Badge,
-    Button,
-    Card,
-    Flex,
-    Spacing,
-    SwitchInput,
-    Text,
-    TextInput,
-    toast,
-    useForm,
-    useService,
-    useStore
-} from "@code0-tech/pictor";
+import React from "react";
+import {useService, useStore} from "@code0-tech/pictor";
 import {UserService} from "@edition/user/services/User.service";
 import {notFound} from "next/navigation";
-import {Tab, TabContent, TabList, TabTrigger} from "@code0-tech/pictor/dist/components/tab/Tab";
-import {IconLayoutSidebar} from "@tabler/icons-react";
-import CardSection from "@code0-tech/pictor/dist/components/card/CardSection";
-import {ApplicationService} from "@edition/application/services/Application.service";
+import {Tab} from "@code0-tech/pictor/dist/components/tab/Tab";
 import {useUserSession} from "@edition/user/hooks/User.session.hook";
 import {
     ResizableHandle,
@@ -28,13 +12,14 @@ import {
     ResizablePanelGroup
 } from "@code0-tech/pictor/dist/components/resizable/Resizable";
 import {SidebarComponent} from "@core/components/SidebarComponent";
+import {ApplicationGeneralSettingsView} from "@ce/application/views/ApplicationGeneralSettingsView";
+import {ApplicationRestrictionsView} from "@ce/application/views/ApplicationRestrictionsView";
+import {ApplicationTabListView} from "@ce/application/views/ApplicationTabListView";
 
 export const ApplicationSettingsPage: React.FC = () => {
 
     const userStore = useStore(UserService)
     const userService = useService(UserService)
-    const applicationService = useService(ApplicationService)
-    const applicationStore = useStore(ApplicationService)
 
     const currentSession = useUserSession()
 
@@ -43,186 +28,22 @@ export const ApplicationSettingsPage: React.FC = () => {
         [userStore, currentSession]
     )
 
-    const application = React.useMemo(
-        () => applicationService.get(),
-        [applicationStore]
-    )
-
     if (currentUser && !currentUser.admin) {
         notFound()
     }
-
-    const initialValues = React.useMemo(
-        () => ({
-            adminStatusVisible: application?.settings?.adminStatusVisible,
-            organizationCreationRestricted: application?.settings?.organizationCreationRestricted,
-            userRegistrationEnabled: application?.settings?.userRegistrationEnabled,
-            legalNoticeUrl: !!application?.settings?.legalNoticeUrl ? application?.settings?.legalNoticeUrl : null,
-            privacyUrl: !!application?.settings?.privacyUrl ? application?.settings?.privacyUrl : null,
-            termsAndConditionsUrl: !!application?.settings?.termsAndConditionsUrl ? application?.settings?.termsAndConditionsUrl : null,
-        }),
-        [application]
-    )
-
-    const [inputs, validate] = useForm({
-        useInitialValidation: false,
-        initialValues: initialValues,
-        validate: {},
-        onSubmit: (values) => {
-            startTransition(() => {
-                applicationService.applicationUpdate({
-                    adminStatusVisible: values.adminStatusVisible,
-                    organizationCreationRestricted: values.organizationCreationRestricted,
-                    userRegistrationEnabled: values.userRegistrationEnabled,
-                    legalNoticeUrl: !!values.legalNoticeUrl ? values.legalNoticeUrl : null,
-                    privacyUrl: !!values.privacyUrl ? values.privacyUrl : null,
-                    termsAndConditionsUrl: !!values.termsAndConditionsUrl ? values.termsAndConditionsUrl : null,
-                }).then(payload => {
-                    if ((payload?.errors?.length ?? 0) <= 0) {
-                        toast({
-                            title: "The application was successfully updated.",
-                            color: "success",
-                            dismissible: true,
-                        })
-                    }
-                })
-            })
-        }
-    })
 
     return <Tab orientation={"vertical"} defaultValue={"general"} h={"100%"}>
         <ResizablePanelGroup>
             <SidebarComponent id={"1"}
                               title={"Application settings"}
                               description={"General settings and restrictions for your Sculptor application. These settings affect all users and organizations within the application."}>
-                <TabList>
-                    <TabTrigger value={"general"} w={"100%"} asChild>
-                        <Button paddingSize={"xxs"} variant={"none"} justify={"start"}>
-                            <Text size={"md"}>General</Text>
-                        </Button>
-                    </TabTrigger>
-                    <TabTrigger value={"restrictions"} w={"100%"} asChild>
-                        <Button paddingSize={"xxs"} variant={"none"} justify={"start"}>
-                            <Text size={"md"}>Restrictions</Text>
-                        </Button>
-                    </TabTrigger>
-                </TabList>
+                <ApplicationTabListView/>
             </SidebarComponent>
             <ResizableHandle/>
             <ResizablePanel id={"2"} color={"primary"} p={1} style={{borderTopLeftRadius: "1rem"}}>
                 <>
-                    <TabContent value={"general"}>
-                        <Text size={"xl"} hierarchy={"primary"}>General</Text>
-                        <Spacing spacing={"xl"}/>
-                        <Card color={"secondary"}>
-                            <CardSection border>
-                                <Flex justify={"space-between"} align={"center"}>
-                                    <Flex style={{gap: ".35rem", flexDirection: "column"}}>
-                                        <Text size={"md"} hierarchy={"primary"}>Sculptor version</Text>
-                                        <Text size={"md"} hierarchy={"tertiary"}>Version of this application</Text>
-                                    </Flex>
-                                    <Badge color={"info"}>{process.env.NEXT_PUBLIC_SCULPTOR_VERSION}</Badge>
-                                </Flex>
-                            </CardSection>
-
-                            <CardSection border>
-                                <Flex justify={"space-between"} align={"center"}>
-                                    <Flex style={{gap: ".35rem", flexDirection: "column"}}>
-                                        <Text size={"md"} hierarchy={"primary"}>Pictor version</Text>
-                                        <Text size={"md"} hierarchy={"tertiary"}>Version of the UI component
-                                            library</Text>
-                                    </Flex>
-                                    <Badge
-                                        color={"info"}>{process.env.NEXT_PUBLIC_PICTOR_VERSION}</Badge>
-                                </Flex>
-                            </CardSection>
-                            <CardSection border>
-                                <Flex justify={"space-between"} align={"center"}>
-                                    <Flex style={{gap: ".35rem", flexDirection: "column"}}>
-                                        <Text size={"md"} hierarchy={"primary"}>Sagittarius version</Text>
-                                        <Text size={"md"} hierarchy={"tertiary"}>Version of the backend</Text>
-                                    </Flex>
-                                    <Badge color={"info"}>{application?.metadata?.version}</Badge>
-                                </Flex>
-                            </CardSection>
-                        </Card>
-                        <Spacing spacing={"xl"}/>
-                        <Flex justify={"space-between"} align={"center"}>
-                            <Text size={"lg"} hierarchy={"primary"}>Legal url's</Text>
-                            <Button color={"success"} onClick={validate}>Save</Button>
-                        </Flex>
-                        <Spacing spacing={"xl"}/>
-                        <Card color={"secondary"}>
-                            <CardSection border>
-                                <Flex justify={"space-between"} align={"center"}>
-                                    <Flex style={{gap: ".35rem", flexDirection: "column"}}>
-                                        <Text size={"md"} hierarchy={"primary"}>Legal notice url</Text>
-                                    </Flex>
-                                    <TextInput {...inputs.getInputProps("legalNoticeUrl")}/>
-                                </Flex>
-                            </CardSection>
-                            <CardSection border>
-                                <Flex justify={"space-between"} align={"center"}>
-                                    <Flex style={{gap: ".35rem", flexDirection: "column"}}>
-                                        <Text size={"md"} hierarchy={"primary"}>Privacy information url</Text>
-                                    </Flex>
-                                    <TextInput {...inputs.getInputProps("privacyUrl")}/>
-                                </Flex>
-                            </CardSection>
-                            <CardSection border>
-                                <Flex justify={"space-between"} align={"center"}>
-                                    <Flex style={{gap: ".35rem", flexDirection: "column"}}>
-                                        <Text size={"md"} hierarchy={"primary"}>Terms and conditions url</Text>
-                                    </Flex>
-                                    <TextInput {...inputs.getInputProps("termsAndConditionsUrl")}/>
-                                </Flex>
-                            </CardSection>
-                        </Card>
-                    </TabContent>
-                    <TabContent value={"restrictions"}>
-                        <Flex justify={"space-between"} align={"center"}>
-                            <Text size={"xl"} hierarchy={"primary"}>Restrictions</Text>
-                            <Button color={"success"} onClick={validate}>Save</Button>
-                        </Flex>
-                        <Spacing spacing={"xl"}/>
-                        <Card color={"secondary"}>
-                            <CardSection border>
-                                <Flex justify={"space-between"} align={"center"}>
-                                    <Flex style={{gap: ".35rem", flexDirection: "column"}}>
-                                        <Text size={"md"} hierarchy={"primary"}>Organization creation</Text>
-                                        <Text size={"md"} hierarchy={"tertiary"}>Set if organization creation is
-                                            restricted
-                                            to
-                                            administrators.</Text>
-                                    </Flex>
-                                    <SwitchInput w={"40px"}
-                                                 {...inputs.getInputProps("organizationCreationRestricted")}/>
-                                </Flex>
-                            </CardSection>
-                            <CardSection border>
-                                <Flex justify={"space-between"} align={"center"}>
-                                    <Flex style={{gap: ".35rem", flexDirection: "column"}}>
-                                        <Text size={"md"} hierarchy={"primary"}>User registration</Text>
-                                        <Text size={"md"} hierarchy={"tertiary"}>Set if user registration is
-                                            enabled.</Text>
-                                    </Flex>
-                                    <SwitchInput w={"40px"}
-                                                 {...inputs.getInputProps("userRegistrationEnabled")}/>
-                                </Flex>
-                            </CardSection>
-                            <CardSection border>
-                                <Flex justify={"space-between"} align={"center"}>
-                                    <Flex style={{gap: ".35rem", flexDirection: "column"}}>
-                                        <Text size={"md"} hierarchy={"primary"}>Admin status</Text>
-                                        <Text size={"md"} hierarchy={"tertiary"}>Set if users can se who is admin</Text>
-                                    </Flex>
-                                    <SwitchInput w={"40px"}
-                                                 {...inputs.getInputProps("adminStatusVisible")}/>
-                                </Flex>
-                            </CardSection>
-                        </Card>
-
-                    </TabContent>
+                    <ApplicationGeneralSettingsView/>
+                    <ApplicationRestrictionsView/>
                 </>
             </ResizablePanel>
         </ResizablePanelGroup>
